main/frcnn_processor/requirements.txt
torch
torchvision
pandas
numpy
# opencv-python
Pillow
urllib3<2.0